requirements.txt
asn1crypto==0.24.0
cached-property==1.5.1
certifi==2019.3.9
cffi==1.12.3
cryptography==2.6.1
Django==2.2.1
django-configurations==2.1
django-telegrambot==1.0.1
future==0.17.1
psycopg2==2.8.2
pycparser==2.19
python-telegram-bot==11.1.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
